Include exported struct methods
This analyzer is great at supporting Builders where the builder is an interface and you're trying to clone an object via that builder. If that builder is a concrete type, that same pattern doesn't work though, because we're only taking methods from interface types. Include methods on concrete structs as well, so we can catch those kinds of problems!